:sparkles: Feature: router 파일 변경
에러경로 추가
